Restructure tests
Use integration tests for Context dependent stuff
Drop roboelectric
