+++ app/src/sharedTest/resources/expected_output_of_populated_book_test.json
